Move Backdraft to talents folder, because it is a talent
